+++ NEWS.md
@@ -1,3 +1,8 @@
+# ukbabynames 0.3.0
+* Added new author in DESCRIPTION.
+* Updated `nibabynames`, `ewbabynames` with 2020 data.
+* Updated tests for added data.
+
 # ukbabynames 0.2.0
 
 * Added new authors and changed maintainer in DESCRIPTION.